reorganize slack event handler to slack folder
